ASoC: msm8x16: update rx path register write sequence
The present code at times disables the pdm clock before
the complete sequence is powered down. The code was designed
in such a way so as to avoid dac being disabled after disabling
the pdm clock.
Alternatively, update the dapm supply widgets by giving
them subsequence id so that the supply widgets are getting
powered down in the correct order to avoid such instances.

CRs-Fixed: 730111
Change-Id: I07f3347b66823633d842f5bea2cd8e0526defa9e
Signed-off-by: Aravind Kumar <akumark@codeaurora.org>
Signed-off-by: Satish Kamuju <skamuj@codeaurora.org>

Conflicts:
	sound/soc/codecs/msm8x16-wcd.c